Migrate toDollarString over to utils/formatters.js
